select: define fd_mask
fd_mask is the type of elements of an fd_set. Searching around, it appears that
this is in the POSIX standard, although I don't have it to hand. (The POSIX
standard also requires that fd_set elements be signed longs, so we certainly
aren't perfect.) Code search suggests that lots of code uses it also.

LSB defines it:
  http://refspecs.freestandards.org/LSB_3.1.0/LSB-Core-generic/LSB-Core-generic.html

The immediate benefit to defining it is that libevent requires it to build.