Update patch set 1
No score.

Oh, Yes. Just checked the gensyscalls.py and that syntax in SYSCALL.txt
is only for socketcall. So I changed the implementation like glibc does.
I added a single ipc syscall for x86 in SYSCALL.txt and implemented a
set of wrappers to convert those sem/shm/msg calls to ipc syscall. Now
it works correctly. 
I have not updated the patch for one question about the declaration of
semctl function. It uses a variable number of arguments. Why can not it
use "union semun" as the 4th argument directly? I changed that in my
latest modification and did not see problems on x86. But not sure about
ARM. 

Thanks
Jackie

Patch-set: 1