Update patch set 1
Patch Set 1: No score

Increase the number of shared libs (.so) that could be open
    
Increasing the number up to 128.
    
This is not a final solution for the limit of shared lib    that could be open.
    
This is a workaround to be able to continue developing while waiting for a final solution from Google and 3rd parties providers to reduce the number of libs been loaded.

Patch-set: 1
Reviewer: Gerrit User 1001586 <1001586@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0