Update patch set 1
On x86 build, bionic implements llrint (in src/s_llrint.c) by calling rint.  It looks perfectly fine from the source code.  But when gcc compiles it, the call to rint is replaced by a call to llrint itself.  When an application calls llrint, it simply crashes because of infinite function recursion.

This example is x86 only and -fno-builtin guarantees correctness.  But it may also slow down libm.  Any suggestion?

Patch-set: 1