stdio: on setvbuf(), move NULL check to where it makes sense and can occur
In the current position, we would check for NULL in a situation
where we are sure that buf would sometimes not be NULL (i.e. when
malloc() succeeded). So, we move the NULL check to the position where
buf can possibly be NULL.

This avoids an extra check when malloc() succeeds, which should be the
most probable case.

Signed-off-by: André Goddard Rosa <andre.goddard@gmail.com>