[PATCH] Fix for sizeof taken on array instead of first element of array  in sysconf.c
The function line_parser_getc (LineParser* p) read 128 characters of
input for each invocation, but increased the position pointer in the
buffer by only one character. This caused the lines to be read to be
sampled more than parsed.

This was caused by taking the sizeof of p->input instead of the first
element of the array (p->input[0]). This patch fixes that.

Change-Id: I93eec3c8c9b9f19ef798748579d0977111b5c0bb