Update patch set 1
Patch Set 1: Do not submit

On second thought, I prefer that we don't submit this at the moment. That's because I'll submit asap another patch corresponding to the state of the internal tree.

The issues are subtle: to properly implement static C++ construction/destruction, we need to define __dso_handle in crtbegin_xxx.c, as well as introduce crtbegin_so.S / crtend_so.S for ARM.

Patch-set: 1
Label: Code-Review=-2