Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

(5 inline comments)

NOTE: something missing here is the fact that dlclose() is supposed to unregister all the handlers that were registered in a shared library when it is unloaded. And there is no official API to unregister the handlers in POSIX.

Solving this problem requires modifying the build system and is pretty hairy (same goes for atexit() handlers registered by a DLL). Just some information to remember for a later patch.

Patch-set: 2
Label: Code-Review=-1
Label: Verified=0