Update patch set 3
Patch Set 3: Looks good to me, approved

I was not asking for you to implement the dlclose() cleanup, just that this is something that we will need to keep in mind.

The current code looks good enough for me. There are a few issues with the fact that the pthread_fork() handlers can in theory call pthread_atfork() and completely corrupt the fork process (since we're using a recursive mutex here), but I think this is a limitation we can live with at the moment.

I would appreciate if you could put a comment regarding it though inside pthread-atfork.c.

Looking at the GLibc implementation, one can see that it does things in a very complex way:

- it doesn't lock the global list when parsing it to call the hooks. Probably to survive pthread_atfork() calls during this phase.

- the dlclose() cleanup is performed by putting the pthread_atfork() implementation inside /usr/lib/libc_nonshared.a (just like atexit()). This means that each shared library has a copy of this function (and atexit() if used), but they can easily pass the address of the __dso_handle to __cxa_atexit() and equivalent. This relies on a specific implementation of weak dynamic linking and I'm not sure that this is something we want to do. Maybe getting the caller address should be enough, but may introduce weird C library <-> dynamic linker interface issues.

I think we will be able to deal with this in the future. In the meantime, the code is good enough. As said before, I'd appreciate a comment.

Also, can you update docs/CHANGES.TXT ?

Patch-set: 3
Label: Code-Review=+2
Label: Verified=0