Update patch set 3
Patch Set 3:

Sorry, I posted that comment before reading your latest.

I deliberately made it a recursive mutex to prevent a deadlock in that situation.  I haven't used the sys/queue.h structures before, but they look well-behaved enough to survive somebody munging up the list during the middle of iterating it, especially since the only operation that can happen is appending a new item to the end of the list.  If so, I think we're good even if somebody calls pthread_atfork() from within a handler.  Is that a correct assessment, or is there a risk there that I'm missing?

As far as glibc's implementation of the __dso_handle stuff, I agree it's pretty arcane, but that's glibc for you.  Since bionic doesn't have a statically-linked component like that, it would be pretty invasive to implement it that way.  I do think the __builtin_return_address() thing would work, though, even though it always gives me the heebie-jeebies to do that.  It would come down to whether you wanted to introduce a secret linker function or not, since I'm pretty sure that would be required to retrieve the __dso_handle.

In any case, I'll table that for now and focus on just getting this part in.  I would appreciate comments on my first paragraph, though--I'm not sure I understand why there's a problem there.

Patch-set: 3