Update patch set 1
Patch Set 1:

It looks like we (Google) have been making similar changes in our internal tree, though not quite identical: I see that "our" second parameter for pwrite is a const void*, whereas "yours" is a void*. I'm not an expert, so I don't know which of the two is most desirable.

Digit, because there are conflicts when merging those changes into our internal master, I'm not going to mark it as Verified myself. Digit, if you're comfortable merging this change (i.e. dealing with the conflicts that will arise), please mark it as Verified and submit it.

Patch-set: 1