Update patch set 1
Patch Set 1: Abandoned

The root cause has been identified.

The IA toolchain is placing the initializer pointers in .ctors instead of .init_array and friends. Bionic expects them to be in .init_array. With the toolchain fixed, this commit is unnecessary. A fix for the toolchain (gcc 4.4) will be submitted upstream and to Google once the fix is accepted into the tip.

Patch-set: 1