Update patch set 1
Patch Set 1:

Well, good news everyone, someone on our side decided that relying on uninitialized data was a bug worth fixing. They found and two places where that was the case already, and that fixes our most visible use case. This might be enough to use the fixed libc.

Patch-set: 1