Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

The first part of your commit wasn't missed. The change is pending... I believe that the fenv.h change is about ready to merge. The link for that one is:
https://review.source.android.com/17765


The libm change didn't pop up with the 4.5.2 compiler... but I agree that it's a bug. I fixed the related one in that same file and committed it as https://review.source.android.com/#change,17766
The libm fix should go in and should have a more appropriate title (such as "use proper variable typing").

Patch-set: 1
Label: Code-Review=-1