Update patch set 1
Patch Set 1:

Hi,

This is now replaced with https://review.source.android.com//#change,20160
  (now without the copyright notice)

For some reason I failed to replace it the proper way this time. Repo complained about --replace is no longer valid, hence I simply did repo upload (with same change-id in the commit message) but it became a new patch set and not a replacement.

Patch-set: 1