Update patch set 2
Patch Set 2:

quick question about this patch: does the code generated by the compiler depends on specific flags?

I.e. how does the generated code handles hardware memory-barrier instructions? For ARM, an ARMv7 CPU can optionally support them, depending on the way it is designed. There is no way for the code to know in advance on what kind of device it will run, so it would essentially call a kernel-helper function or something like that.

Oh yeah, I don't know if the code that was deleted was SMP-safe by the way, it's been a long time since I looked at the Intel Architecture Reference Manuals :-)

Apart from that, LGTM, but I would like a clarification before accepting it. Thanks for your understanding.

Patch-set: 2