Use environment variable to set dns cache size
Use the environement variable ANDROID_DNS_CACHE_SIZE
to set the max number of entries in the cache. If
ANDROID_DNS_CACHE_SIZE is not set the default value is
defined by CONFIG_MAX_ENTRIES.

The number of entries in the hash table will be equal
to the number of max entries allowed in the cache.

Change-Id: I77d69d7c178937fa483d0b40512483ad29232d28