Wrong return value if __pthread_clone() fails
When creating a native thread using pthread_create() the
__pthread_clone() might return an error code if the new thread
could not be created and it is this error code that must be returned.
This fix stores the errno before calling the munmap() syscall thus
reflecting the actual error triggered by __pthread_clone() rather
than the errno generated by munmap() if any. If munmap() succeedes
the errno will be 0 overriding the previous error code.

Change-Id: I0c43f3e21bbb9385107faf0a8d3faf19d4ae1ac3
Signed-off-by: Oskar Andero <oskar.andero@sonyericsson.com>