Make sure __u64 is defined even for strict ansi or -std=c99
The x86 asm headers define __u64 regardless of __STRICT_ANSI__.
The linux/videodev2.h header requires __u64 to be defined, thus
this fixes compiling with -std=c99 when including the
linux/videodev2.h header.

This line is copied unmodified from the original kernel
header though, I'm not sure if this is the right way of fixing
it, if the header in external/kernel-headers/original should
be modified accordingly, or if the cleanup scripts could be
modified to achieve the same effect?

In glibc, the asm/types.h header defines __u64 regardless of
__STRICT_ANSI__.

Change-Id: If05fcc9ed6ff5943602be121c7be140116e361fe