Set default TTL if response code indicates an error.
The response code (rcode) is set as part of the
response. If the response code  is 0 everything is ok
any other value indicates something went wrong.

Reponses with a response code indicating an error
isn't cached today, which result in unwanted traffic
in the network. This patch solves the problem by
caching these responses a default time i.e. 10 minutes.

The default time is the same time as the cache used before it
started to honor the TTL of the answer records.

Change-Id: I97b84f139755dc4b9cf2670f77c19ad7161552be