bionic: fix incorrect logb implementation when use the subnormal as input
When the subnorml is used as input to logb(), the result is incorrect.
For example, if we use the mininum submornal as the input:
 0x0000000000000001 = 2^(-1022 - 52)
logb(0x0000000000000001) should be -1022 - 52 = -1074.
But we got -1077 by the current logb implementation in bionic/libm.

The root issue is: the high 32bits are not extracted correctly after the
subnormal is converted to the normal by multiplying 2^54.

Fixed it in this patch.

Change-Id: I54f33648db7c421b06eee1ea8e63c57a179fae0d
Signed-off-by: Jack Ren <jack.ren@intel.com>
Signed-off-by: Jin Wei <wei.a.jin@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>