Update patch set 1
Patch Set 1: I would prefer that you didn't submit this; Verified

I believe that the code is indeed equivalent, but the comments are non-obvious at best, and possibly misleading. I think that the two comments need to be merged into one before the loop, explaining why it's unnecessary to clear the top 4 bits within the loop, as long as they're cleared at the end.

Also, this optimization probably fails if ints aren't 32-bit, and it'd be good to at least mention that and possibly to enforce it with a proper fixed-length type.

Patch-set: 1
Label: Code-Review=-1