libc: optimize visibility of system call stubs
This patch modifies the tools/gensyscalls.py script to make all
generated system call stubs which names begins with an underscore
with hidden visibility (i.e. they won't be exported by libc.so).

This convention is easier than adding new syntax/flags in the
SYSCALLS.TXT file format.

There are however a few cases where the platform expects to be
able to link against said functions, so stubs are provided to
maintain the ABI. The complete list of stubs here is:

- __mmap2() is not declared in <sys/mmap> but used by
  system/extras/ext4_utils

- __reboot() is defined by <sys/reboot.h> and used by libcutils
  and some vendor-specific flasher/bootloader.

- __wait4() is not defined by <sys/wait.h> but used by strace.
  Note that this change renames the system call to 'wait4' which
  is declared by <sys/wait.h> but wasn't implemented previously.

- _exit is renamed to __exit_group (which better matches
  the system call number), and a stub _exit() function is
  provided under libc/unistd/exit.c

+ use the ENTRY/ENTRY_PRIVATE/END assembly macros in x86 stubs

+ remove the dependency on <sys/linux-syscalls.h> in the
  generated assembly fragments (instead, hard-code the
  corresponding macro definition inside the generated
  file)

Change-Id: If3519f0aaf236932b650068de247524c0341213a