fix definitions of size_t, ssize_t and ptrdiff_t to be what GCC expects
This fixes warnings of the form:
../src/shf.c:718: warning: format '%zd' expects type 'signed size_t', but argument 3 has type 'ssize_t'

See also commit b11955f1fa7b092e12840699434318f679a2038f in pkg-dietlibc:
http://anonscm.debian.org/gitweb/?p=collab-maint/dietlibc.git;a=blob;f=debian/patches/0028-fix-ssize_t.diff;h=33a5e3f89add5269f9bfd7e5d744c7c5fd791366;hb=b11955f1fa7b092e12840699434318f679a2038f

The fix originates from MirBSD, has been in use there for years,
was recently applied (and tested) to dietlibc, and seems to be
needed here as well. Its only downside is that the headers now
require GCC or a GCC-compatible compiler.

Change-Id: I8866ff3a7e0b468dd72d3306b9987dba4f6cff98