Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

Checking gsm.pdpprotocol here seems like a layering violation - as Robert says, this code could be useful on other types of networks as well (e.g., IPv6-only wifi).

Perhaps what we need is a property that says "we don't have IPv4, but only translated IPv4". Robert, if that property's name was not called gsm.pdpprotocol.ipv6 but something like "net.ipv4.compat" or "net.ipv4.isnative" would it be better?

Also, I'm not sure that this is the right thing to do. This is not only a performance optimization, but it also has effects on correctness. For example, suppose you do a DNS lookup for a hostname that has both an IPv4 address and a 6to4 address. Without this change the app will (correctly, I think) prefer IPv4 NAT to the IPv4 address, but with the change it will prefer IPv6 to the 6to4 address, which will probably be less reliable. Dan, what's your take on this?

Patch-set: 2