Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

Robert brought up the case of having multiple connections active.  For example, a v4-only PDP for something like MMS and v6-only PDP for the general internet, or some day in the future, having both wifi and lte/wimax active at the same time.  This patch does not handle that case, and I'm not able to conceive of a simple patch that does not either:

1. generates surprises when multiple connections are active - bringing up an unrelated v4 connection would change what DNS answers all applications recieved

-or-

2. breaks applications on the secondary connection - the v4 MMS example would not be able to look up the v4 mms address

At this point, I think the solution has gotten more complex than the problem, which was only an marginal optimization to begin with.  Barring any surprises, I'm going to consider this patch dead.

Patch-set: 2
Label: Code-Review=-1