Update patch set 2
Patch Set 2:

Hey Dan, I want to take another look at this.  Lorenzo and I spoke last night and while what you've got has problems from the multi-connection standpoint it's no worse than what we've currently got in that regard.  When we fix the rest of the resolver up to be multi-connection aware we could fix this too.

Patch-set: 2