Update patch set 1
Patch Set 1: Do not submit

Please do not submit this patch to the AOSP tree, it will fail to merge with the internal one.

For the record, we solved the problem very differently in the internal tree (i.e. we completely got rid of the out-of-band mutex hash table). I'm surprised this isn't in AOSP yet, but it will be part of the next ICS open-source release.

Patch-set: 1