Update patch set 2
Patch Set 2:

oops, bad format. re-pasted...

Hi H.J.,
That problematic code is not calling __pthread_create( ), it is in "jmp __thread_entry".
__thread_entry( ) is a standard C function. In order to have its stack be aligned with 16 byte properly, it needs the stack has following layout when it is called:
layout #1:
--------------
|            |
-------------- <--ESP (ECX - 20)
| ret EIP    |
-------------- <--ECX - 16
| arg0       |
-------------- <--ECX - 12
| arg1       |
-------------- <--ECX - 8
| arg2       |
-------------- <--ECX - 4
| unused     |
-------------- <--ECX (16-byte boundary)

But in the current code, its layout is as follows when branching to it: layout #2:
--------------
|            |
-------------- <--ESP (ECX - 16)
| unused     |
-------------- <--ECX - 12
| arg0       |
-------------- <--ECX - 8
| arg1       |
-------------- <--ECX - 4
| arg2       |
-------------- <--ECX (16-byte boundary)

This patch is to make sure the __thread_entry( ) has the layout #1

Patch-set: 2