libstdc++: Fix x86 thread-safe one-time-construction implementation.
The root of the problem is that the existing implementation is based on the
ARM C++ ABI, which mandates that only the lsb of the guard variable should
be tested by compiler-generated code before calling __cxa_guard_acquire et al.

The Itanium / x86 C++ ABIs define a different convention where the low-order
_byte_ is tested instead.

Meanwhile, guard variables are 32-bit aligned for ARM, and 64-bit for x86.

Reference documentation:

    section 3.2.3 of ARM IHI 0041C (for ARM)
    section 3.3.2 of the Itanium C++ ABI specification v1.83 (for x86).

This patch modifies the implementation in a way that satisfies both ABIs (and
doesn't require changing the toolchains).

Change-Id: I885e9adc7f088b9c0a78355bd752f1e6aeec9f07
Signed-off-by: Fengwei Yin <fengwei.yin@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>