bionic: implement tgkill to replace kill
- According to man page:
In a multithreaded program system call raise is equivalent to
pthread_kill(pthread_self(), sig);

but in the implementation, it uses system call kill, and kill sends
signal to a process, not a thread, so it will update shared pending
table in kernel signal descriptor(task->signal->shared_pending),
not the private pending one in kernel(task->pending).thus this pending
signal will be shared by all group threads.

- And tkill is obselete, so replace it with tgkill in the implementaion
of pthread_kill.

Change-Id: I7868ba14c5f0771452c668596d6c705708a09c02
Signed-off-by: Jin Wei <wei.a.jin@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Signed-off-by: Fengwei Yin <fengwei.yin@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>