Update patch set 3
Patch Set 3: (5 inline comments)

Thanks for catching this, and sorry for the late review, I've been very sick lately.

I confirm that the patch will fix the issue properly, but I have added a few comments to explain that I'd like to see a few things rephrased.

I also feel sorry, because I already did handle this for the NDK GABI++ implementation of the same source file (I had to re-implement it to use pthread mutexes instead of futexes for portability reasons), but forgot to back-port the changes to the system.

Patch-set: 3
Label: Code-Review=0
Label: Verified=0