Update patch set 2
Patch Set 2:

okay, this is your change with the PIC change merged in, plus i've extended the PIC changes to cover the new code. it _builds_ but i don't have a working x86 tree to _test_ it. please cherry-pick and test this yourselves, and i'll submit it when you tell me it works.

Patch-set: 2