mallopt() parameters choice and size allocation pattern profiling support
A proposal to let Apps be able to choose its best mallopt() parameters in
order to be more memory efficient and less latency consumer, able to
know its size allocation pattern to devise sub-allocator if need be.

A new file dlmalloc_profile.c is inserted with mallopt() and size profiling
changes, as Daklvik heap uses dlmalloc.c through mspace.c, so this is just
to over simply it, though in future this could be a shared lib similar to
libc_malloc_debug_leak.so to be loaded at startup.
An extra lock for latency calculation is added which would impact the performace
may be removed at cost of tweaking dlmalloc core functions.

Change-Id: Iae04671cac854953a9fdcbde04b716ac4c7595c9