Allow clang to use builtin for ffs()
gcc is already using a builtin for ffs in most cases where it's used
even without having to include strings.h. Clang also has a builtin ffs()
but it's named __builtin_ffs(). This patch enables clang to replace use
of ffs() with its builtin version as long as the project includes
strings.h

Change-Id: I1bfad001f107fff2758d3e6cfc58051f73660849
Author: Edwin Vane <edwin.vane@intel.com>
Reviewed-by: Kevin P Schoedel <kevin.p.schoedel@intel.com>