strtod: Support parsing hex integers
This adds support for a bit more of what the standard mandates.
This doesn't implement full support for parsing hex floats with
an exponent though.

Support for hex strings in strtod in the NetBSD implementation
came as part of a larger overhaul (where most of the parsing
routines were moved to a separate subdirectory "gdtoa") - importing
that into bionic would be a much bigger change.

Change-Id: If7a6085f3d8336c246351b2c4da3fc2166fa0b58