Update patch set 1
Patch Set 1:

(sorry about the delay).

It's probably not strictly necessary to relicense those, and given that one of the files comes from ARM that could end up taking a long time.

Elliott, if this is fine with you, it works for me, but we'll need the NOTICE file updated accordingly.

Patch-set: 1