Update patch set 2
Patch Set 2:

Sorry if I'm beating a dead horse here, but since this patch hasn't seen the light of day yet I figure it is still worth improving it if it makes sense.

The .note.ABI-tag I added in this patch uses the GNU vendor namespace. I learned that as such it is reserved for GNU tools use and the meaning and layout of notes using the GNU name is coordinated between these tools. Roland McGrath would rather have an Android flag in a note in a different namespace instead of extending the GNU ABI-tag.

So I'm thinking of creating .note.android.ident, using the "Android" name and having a single 32 bit value as contents, containing the API level the app expects. NetBSD and OpenBSD have .note.netbsd.ident and .note.openbsd.ident so it's consistent with existing usage.

What do you think? Would you mind if I push a third version of the patch with that change?

Patch-set: 2