linker: Fix relocation of R_ARM_COPY type relocations.
For the case of R_ARM_COPY, it makes no sense to do a local lookup for
the symbol, as it will find the very entry it is supposed to relocate to,
which in turn means that the memcpy() would copy the data over itself,
doing nothing useful.

Add a ignore_Local parameter to _do_lookup, which is used in this special
case to ignore symbols in the local soinfo.