Update patch set 1
Patch Set 1:

i see. yeah, last time we fiddled with this check the consensus was that since we know the ARM implementation always copies forwards, the extra test was unnecessary. i don't remember which side of the argument i was on.

i can guess, though, given that i'm the kind of person who'd be happy aliasing memcpy to memmove.

do the systems team benchmark memmove or just memcpy? will they notice this change?

Patch-set: 1