Update patch set 2
Patch Set 2:

i guess the bit i don't get is when do we end up in the runtime check and what does it mean? the compile-time checks make sense, but i'm having trouble understanding the utility of the runtime check. it seems like it would only cover "crazy" code like your example in the comment.

Patch-set: 2