Update patch set 2
Patch Set 2:

My only question is does this new NOTICE content making it to anywhere. If the NOTICE file is not in the same directory as the LOCAL_PATH when doing an "include $(BUILD_...)" I do not think it will use the NOTICE file in any generated output.

I just confirmed with enh that this is the case with his change. For example, he doesn't see something like this anymore:

Notice file: bionic/libc/NOTICE -- out/target/product/maguro/obj/NOTICE_FILES/src//system/lib/libc.so.txt

(I realize the source will change, but not the destination as I understand it)

Patch-set: 2
Label: Code-Review=0