Update patch set 1
Patch Set 1:

I couldn't find the implementation/invocation of _bionic_heap_error. Is it replaced by something else? Also, in 2.8.3 we modified CORRUPTION_ERROR_ACTION and USAGE_ERROR_ACTION to pass the bad heap pointer to _bionic_heap_error so that the nearby memory content can be dumped in the tombstone file. It is very valuable to debug heap corruptions so I hope we can preserve that.

Patch-set: 1
Label: Code-Review=0