Update patch set 2
Patch Set 2:

PS2 addresses review comments from PS1. It adds a usage and heap corruption error handler. The details in the message are changed slightly from the current handler to make it more obvious the corruption lies in libc and that an argument is invalid rather than general corruption (when that is the case).

Patch-set: 2
Reviewer: Gerrit User 1010118 <1010118@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0