ARM: fixup atexit() calls from legacy ARM shared libraries
On other archs, exit functions of shared libraries are
correctly scheduled for execution at dlclose(). This is
implemented by putting a private copy of atexit() in the
CRT startup files.

On ARM, this behaviour is corrected by the previous patch,
but existing NDK built shared libraries will still call
the generic atexit() supplied by libc.so. This patch adds
a workaround to that version of atexit() that results in
the exit functions being scheduled correctly.

Change-Id: I41ae153c23268daa65ede7fb8966fc3e9caec369
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@gmail.com>