Update patch set 3
Patch Set 3:

Ok, I just checked a few things, and I noticed a small problem.

In theory, the NDK-provided libc.so does not expose atexit() or __dso_handle (libc.so is an empty ELF file generated from a list of symbols).

I say "in theory" because I was convinced I had modified the scripts (e.g. $NDK/build/tools/gen-symbols.sh) to do that for all platforms. It seems this isn't the case though.

I don't know why, but this is definitely an NDK bug, and I'll provide a fix soon. Will update this entry soon.

Patch-set: 3
Label: Verified=0