Update patch set 3
Patch Set 3:

Ok, I've uploaded a fix for the NDK issue at https://android-review.googlesource.com/#/c/42241/3

Interestingly, I discovered by running the new unit test that the MIPS C library did expose "atexit", but not __dso_handle. This certainly is an error in the way the library is built.

Given that MIPS support integration to the platform is not completed, I'm in favor of removing this from the C library once the NDK fix is in, but I'll let Elliot decide on this.

Patch-set: 3