Update patch set 1
Patch Set 1:

I'm not really sure why ffs() got singled out, but Clang seems to have few (close to zero) builtin functions that are not prefixed with __builtin_. I don't see them changing this. 

Edwin, it actually looks like this is a build system issue, where we are not building arch-x86/string/fss.S (sic). That has a baseline implementation of ffs() too, which might be the actual fix here.

Patch-set: 1